# openapi-codegen/openapi.yaml
openapi: 3.0.1
info:
  title: Task Management API
  version: 1.0.0
  description: API for managing tasks in the Task Management System.

paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all tasks.
      operationId: getTasks
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      description: Add a new task to the system.
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: Task created successfully.

  /tasks/{id}:
    get:
      summary: Get a task by ID
      description: Retrieve a specific task by its ID.
      operationId: getTaskById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found.
    put:
      summary: Update a task
      description: Modify the details of an existing task.
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task updated successfully.
    delete:
      summary: Delete a task
      description: Remove a task from the system.
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully.

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - IN_PROGRESS
            - COMPLETED
        priority:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH